Appointment Scheduling improvements
After submitting the form to search for available slots, all the
fields get disabled, making it impossible to refined the search
by changing one or two params. The only option available is to
start the search all over again, which is a bit annoying.

This fix will automatically enabled the fields again after the
search is done, so that it's possible to change what is necessary
without having to start again